pyproject.toml
[tool.poetry]
name = "pet-walking"
version = "0.1.0"
description = ""
authors = ["KondratevProkhor <prokhor-2000@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.1"}
tortoise-orm = "^0.21.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
